organize files add animation effects
