+++ src/page1/css/modal.css
@@ -13,25 +13,57 @@
 /* модальноые окна с вопросами*/
 
 .modal__btn{
-  padding: 10px 20px;
+  padding: 15px 20px;
   width: 10%;
-  margin: 20px;
+  margin: 40px;
  
   
-  color: rgb(22, 22, 22);
+  
   font-size: 1.25em;
   border: none;
   border-radius: 2em;
 
   background: #f8f8f8;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25), inset 2px 2px 4px rgba(0, 0, 0, 0.08);
+  box-shadow: 0px 4px 4px #f8f8f8(0, 0, 0, 0.25), inset 2px 2px 4px rgba(0, 0, 0, 0.08);
   
 
 
   margin: 0;
   cursor: pointer;
 
+  transition: 0.5s;
 
+  
+  
+}
+.modal__btn span{
+  cursor: pointer;
+  display: inline-block;
+  position: relative;
+  transition: 0.5s;
+  font-family: 'Playfair Display SC', serif;
+  font-size: 0.75em;
+  letter-spacing: 0.1em;
+  text-transform: uppercase;
+  color: rgb(22, 22, 22);
+}
+
+/* .modal__btn:hover span {
+  transform: scale(1) translate(-10%);
+} */
+
+
+
+.modal__btn:hover{
+  background: rgb(102, 0, 0);
+  box-shadow: 0 12px 16px 0 rgb(192, 192, 192)(0,0,0,0.24), 0 17px 50px 0 rgba(0,0,0,0.19);
+  width: 12%;
+  
+  
+}
+.modal__btn:hover span {
+  color: white;
+  letter-spacing: 0.2em;
 }
 
 
@@ -48,7 +80,10 @@
 
   visibility: hidden;
   opacity: 0;
-  transition: visibility 0s opacity 0.5s;
+  transition: all 0.8s ease 0s;
+
+  overflow-y: auto;
+  overflow-x: hidden;
 
 
 }
@@ -58,12 +93,20 @@
 
 
 }
+.modal__bg_active .modal{
+  transform: translate(0px, 0px); /* включаем выплываение сверху */
+  opacity: 1;
+}
 .modal{
   position: relative;
   background-color: #f8f8f8;;
   width: 60%;
   height: 50%;
   color: rgb(22, 22, 22);
+
+  transition: all 0.8s ease 0s; /*выплываение окна сверху*/
+  opacity: 0;
+  transform: translate(0px, -100%);
 
 }
