+++ src/page1/index.css
@@ -50,6 +50,8 @@ h1 {
   text-align: center;
   font-size: 1.5em;
   font-weight: bold;
+
+  margin: 0 0 40px 0 ;
 }
 
 h2 {
