+++ src/page1/index.html
@@ -31,7 +31,7 @@
     
     <section class="text"> 
       <h1 >КТО ТЫ ИЗ ИГРЫ ПРЕСТОЛОВ?</h1>
-      <button class="modal__btn">Начать</button>
+      <button class="modal__btn"><span>начать</span></button>
     </section>
 
     <!--Модальные окна -->
