+++ src/page1/index.js
@@ -1,4 +1,4 @@
-// import modal from "./js/modal.js";
+import {openModalInfo,closeModalInfo, openModal, closeModal} from './js/modal.js'; 
 
 
 let radio = new Audio();
@@ -118,34 +118,5 @@ document.querySelector('#off').onclick = function() {
 //     });
 //   });
 
-//модальное окно с инфой
 
-const info = document.querySelector(".img__question");
-const modalBg = document.querySelectorAll('.modal__bg');//nodelist всех модальных окон
-
-info.addEventListener("click", openModalInfo);
-function openModalInfo(){
-    modalBg[0].classList.add("modal__bg_active");
-}
-
-info.addEventListener("mouseout",closeModalInfo);
-
-function closeModalInfo(){
-    modalBg[0].classList.remove("modal__bg_active");
-}
-
-
-// // модальное окно вопросы
-const modalBtn = document.querySelector(".modal__btn");
-const modalCloseBtn = document.querySelector(".modal__close");
-
-modalBtn.addEventListener("click", openModal);
-function openModal(){
-    modalBg[1].classList.add("modal__bg_active");
-}
-
-modalCloseBtn.addEventListener("click", closeModal);
-function closeModal(){
-  modalBg[1].classList.remove("modal__bg_active");
-}
